# openapi.yaml
openapi: 3.0.3
info:
  title: QuoteSlate API
  description: |
    # 🎯 QuoteSlate API - Inspirational Quotes at Your Fingertips
    
    A powerful, developer-friendly REST API providing access to a curated collection of **2,616+ inspirational quotes** from **1,010+ renowned authors** across **31 categories**.
    
    ## ✨ Key Features
    
    - **🎲 Random Access**: Get truly random quotes for inspiration
    - **📄 Systematic Browsing**: Browse quotes with advanced pagination
    - **🔍 Smart Search**: Search quotes by content and author names
    - **👤 Author-Specific**: Get all quotes from specific authors
    - **🏷️ Tag-Based Filtering**: Filter quotes by categories/tags
    - **⚡ Lightning Fast**: Average response time under 10ms
    - **🧠 Smart Caching**: Random endpoints stay random, others are optimized
    - **🔒 Secure**: Comprehensive input validation and security headers
    
    ## 🚀 Getting Started
    
    No API key required! Simply make HTTP requests to any endpoint.
    
    **Base URL**: `https://quoteslate.vercel.app`
    
    ## 📊 Rate Limits
    
    - **100 requests per 15 minutes** per IP address
    - Use the `count` parameter to fetch multiple quotes efficiently
    - Consider caching responses in your application
    
    ## 🎯 Best Practices
    
    1. **Optimize Requests**: Use `count` parameter for bulk fetching
    2. **Cache Responses**: Store quotes locally to reduce API calls
    3. **Handle Errors**: Always check response status codes
...
    
...
servers:
...
paths:
...
  /api/quotes/random:
    get:
      tags:
        - Random Quotes
      summary: Get Random Quote(s)
      description: |
        Retrieve one or more random quotes from the collection. Perfect for inspiration, daily quotes, or adding variety to your application.
        
        **⚠️ Important**: This endpoint is NOT cached to preserve true randomness.
      operationId: getRandomQuotes
      parameters:
        - name: count
          in: query
          description: Number of random quotes to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 1
          example: 5
        - name: authors
          in: query
          description: Comma-separated list of author names to filter by
          required: false
          schema:
            type: string
          example: "Albert Einstein,Maya Angelou"
        - name: tags
          in: query
          description: Comma-separated list of tags to filter by
          required: false
          schema:
            type: string
          example: "motivation,wisdom"
        - name: minLength
          in: query
          description: Minimum character length of quotes
          required: false
          schema:
            type: integer
            minimum: 1
          example: 50
        - name: maxLength
          in: query
          description: Maximum character length of quotes
          required: false
          schema:
            type: integer
            minimum: 1
          example: 150
      responses:
        '200':
          description: Random quote(s) retrieved successfully
          headers:
            Cache-Control:
              description: No cache headers to preserve randomness
              schema:
                type: string
                example: "no-cache, no-store, must-revalidate"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Quote'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Quote'
              examples:
                single_quote:
                  summary: Single random quote
                  value:
                    id: 498
                    quote: "Every strike brings me closer to the next home run."
                    author: "Babe Ruth"
                    length: 51
                    tags: ["wisdom", "motivation"]
                multiple_quotes:
                  summary: Multiple random quotes
                  value:
                    - id: 498
                      quote: "Every strike brings me closer to the next home run."
                      author: "Babe Ruth"
                      length: 51
                      tags: ["wisdom", "motivation"]
                    - id: 120
                      quote: "All great achievements require time."
                      author: "Maya Angelou"
                      length: 36
                      tags: ["motivation", "success"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimit'
    head:
      tags:
        - Random Quotes
      summary: Get Random Quote Headers
      description: Get headers for random quotes endpoint without response body
      operationId: getRandomQuotesHeaders
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Headers retrieved successfully
  /api/quotes:
    get:
      tags:
        - Paginated Quotes
      summary: Browse All Quotes (Paginated)
      description: |
        Browse the entire quote collection systematically with advanced pagination, search, sorting, and filtering capabilities.
        
        **✅ Cached**: This endpoint is cached for 5 minutes for optimal performance.
      operationId: getPaginatedQuotes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - name: search
          in: query
          description: Search term to find in quote text or author names
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
          example: "love"
        - name: authors
          in: query
          description: Comma-separated list of author names to filter by
          required: false
          schema:
            type: string
          example: "Albert Einstein,Buddha"
        - name: tags
          in: query
          description: Comma-separated list of tags to filter by
          required: false
          schema:
            type: string
          example: "motivation,wisdom"
        - name: minLength
          in: query
          description: Minimum character length of quotes
          required: false
          schema:
            type: integer
            minimum: 1
        - name: maxLength
          in: query
          description: Maximum character length of quotes
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Paginated quotes retrieved successfully
          headers:
            Cache-Control:
              description: Cached for performance optimization
              schema:
                type: string
                example: "public, max-age=300, s-maxage=300"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuotesResponse'
              example:
                data:
                  - id: 1
                    quote: "Action is eloquence."
                    author: "William Shakespeare"
                    length: 20
                    tags: ["motivation"]
                  - id: 2
                    quote: "When in doubt, don't."
                    author: "Benjamin Franklin"
                    length: 21
                    tags: ["wisdom"]
                pagination:
                  page: 1
                  limit: 20
                  total: 2616
                  totalPages: 131
                  hasNext: true
                  hasPrev: false
                  startIndex: 1
                  endIndex: 20
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimit'
  /api/quotes/by-author/{author}:
    get:
      tags:
        - Authors
      summary: Get Quotes by Specific Author
      description: |
        Retrieve all quotes from a specific author with pagination and sorting options.
        
        **✅ Cached**: This endpoint is cached for 5 minutes for optimal performance.
      operationId: getQuotesByAuthor
      parameters:
        - name: author
          in: path
          description: Author name (case-insensitive, URL-encoded)
          required: true
          schema:
            type: string
            minLength: 1
          example: "Albert Einstein"
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: sort
          in: query
          description: Sort quotes by field
          required: false
          schema:
            type: string
            enum: [id, length, random]
            default: id
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Author quotes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuotesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Author not found or no quotes available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Author "NonExistent Author" not found.'
        '429':
          $ref: '#/components/responses/RateLimit'
  /api/quotes/by-tag/{tag}:
    get:
      tags:
        - Tags
      summary: Get Quotes by Specific Tag
      description: |
        Retrieve all quotes with a specific tag/category with pagination and sorting options.
        
        **✅ Cached**: This endpoint is cached for 5 minutes for optimal performance.
      operationId: getQuotesByTag
      parameters:
        - name: tag
          in: path
          description: Tag name (case-sensitive, URL-encoded)
          required: true
          schema:
            type: string
            minLength: 1
          example: "motivation"
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Tag quotes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuotesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Tag not found or no quotes available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Tag "nonexistenttag" not found.'
        '429':
          $ref: '#/components/responses/RateLimit'
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
...
  responses:
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Too many requests, please try again later. Consider using the 'count' parameter to fetch multiple quotes in a single request."
  securitySchemes: {}
security: []
...